Move state management to presenter
