+++ starwarswiki/src/views/hamburgerView.jsx
@@ -1,6 +1,5 @@
 import { slide as Menu } from 'react-burger-menu';
 import { Link } from 'react-router-dom';
-import { useState } from 'react';
 import Logo from '../../assets/images/sww-logo.svg';
 import SignInButton from '../components/SignInButton.jsx';
 import GoToProfile from '../components/GoToProfile.jsx';
@@ -8,21 +7,6 @@ import GoToProfile from '../components/GoToProfile.jsx';
 import '../style/burgermenu.scss';
 
 export default function HamburgerView(props) {
-	const [menuOpen, setMenuOpen] = useState(false);
-
-	function handleOnClick(event) {
-		props.onClickHandler(event.target.innerText.toLowerCase());
-		setMenuOpen(!menuOpen);
-	}
-
-	function handleOnOpen() {
-		setMenuOpen(true);
-	}
-
-	function handleOnClose() {
-		setMenuOpen(false);
-	}
-
 	function isMenuOpen(state) {
 		return state.isOpen;
 	}
@@ -33,23 +17,23 @@ export default function HamburgerView(props) {
 				<img src={Logo} alt='Star Wars Wiki logo of green Yoda with red Christmas hat' />
 			</Link>
 			<Menu
-				isOpen={menuOpen}
-				onOpen={handleOnOpen}
-				onClose={handleOnClose}
+				isOpen={props.isOpen}
+				onOpen={props.handleOnOpen}
+				onClose={props.handleOnClose}
 				onStateChange={isMenuOpen}
 				right
 			>
-				<Link to='/characters' onClick={handleOnClick}>
+				<Link to='/characters' onClick={props.handleOnClick}>
 					Characters
 				</Link>
-				<Link to='/vehicles' onClick={handleOnClick}>
+				<Link to='/vehicles' onClick={props.handleOnClick}>
 					Vehicles
 				</Link>
-				<Link to='/locations' onClick={handleOnClick}>
+				<Link to='/locations' onClick={props.handleOnClick}>
 					Locations
 				</Link>
 				{props.user ? (
-					<GoToProfile inAnimation={props.inAnimation} onClickHandler={handleOnClick} />
+					<GoToProfile inAnimation={props.inAnimation} onClickHandler={props.handleOnClick} />
 				) : (
 					<SignInButton />
 				)}
